Feat: CREATE NEW CUSTOM HOOK FOR HANDLE ADD STOCK AND REMOVE IT FROM THE STOCKFORM COMPONENT
